move: actionsAndAttributes navbar to dedicated dir from components
